Prefer smaller movements when solution searching
This changes the search order so instead of fully rotating/shifting
each ring/row, it will "jiggle" each and check for solutions that way
first. Right now it seems to work OK on the main thread when in
release mode, but that may need to change.

I really tried to get the fancy (commented out) iterator setup
going. Unfortunately I had to fall back to callbacks and for loops.
Ownership can be a bit draconian in Rust...